third_party/WebKit/LayoutTests/svg/dom/SVGViewSpec-invalid-ref-crash-expected.txt

CONSOLE WARNING: line 18: SVGSVGElement.currentView is deprecated and will be removed in M56, around January 2017. See https://www.chromestatus.com/features/4511711998509056 for more details.
Confirm that null is returned when an SVGViewSpec is used after its corresponding element has been removed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS svgView.preserveAspectRatio; is non-null.
PASS svgView.transform; is non-null.
PASS svgView.viewTarget; is null
PASS svgView.viewBox; did not throw exception.
PASS svgView.preserveAspectRatioString is "xMidYMid meet"
PASS svgView.transformString is ""
PASS svgView.viewBoxString is "0 0 0 0"
PASS successfullyParsed is true

TEST COMPLETE
